+++ tools/watcher/castor.php
@@ -17,7 +17,8 @@ function linux()
 #[AsTask(description: 'Build watcher for MacOS system')]
 function darwin()
 {
-    run('go build -o bin/watcher-darwin -ldflags="-s -w" main.go', environment: ['GOOS' => 'darwin', 'CGO_ENABLED' => '0'], path: __DIR__);
+    run('go build -o bin/watcher-darwin-amd64 -ldflags="-s -w" main.go', environment: ['GOOS' => 'darwin', 'GOARCH' => 'amd64', 'CGO_ENABLED' => '0'], path: __DIR__);
+    run('go build -o bin/watcher-darwin-arm64 -ldflags="-s -w" main.go', environment: ['GOOS' => 'darwin', 'GOARCH' => 'arm64', 'CGO_ENABLED' => '0'], path: __DIR__);
 }
 
 #[AsTask(description: 'Build watcher for Windows system')]